% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_combination.R
\name{best_n_with_diff}
\alias{best_n_with_diff}
\title{Function that selects a set of bets splitting the selection}
\usage{
best_n_with_diff(data, nsigns, diff)
}
\arguments{
\item{data}{(data.table): A table with expected column sign ordered by
priority of taking (typically em, probability or rentability)}

\item{nsigns}{(integer): number of bets you want to select}

\item{diff}{(integer): min number of signs different between every 2
selected bets.}
}
\value{
selected (data.table): Final bets selected
}
\description{
Normally, if you select the bets by probability, rentability or em,
consecutive bets have just 1 different sign. In order to avoid taking
n very similar bets. This function provides a method to take splited bets
with different signs in different matches.
}
\examples{
library(data.table)
data <- data.table(sign = c('11111111111111', '11111111112222', '11122222111111'))
best_n_with_diff(data, 2, 5)

}
