% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reward.R
\name{calculate_all_reward}
\alias{calculate_all_reward}
\title{Calculate the expected reward for a given prob-voted table}
\usage{
calculate_all_reward(prob_voted_table, money)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table to work with. Watch calculate_probabilities
function for further information.}

\item{money}{(numeric): the total money played by all the bets}
}
\value{
prob_voted_table (data.table): The same prob_voted_table with reward columns added.
}
\description{
The expected reward is the money you will get if you get an specific number
of right matches and the final result is the row sign. For example, if the row has
the sign 111...111 and reward_11 is 20, it means that if the final result is 1
in all matches and you have 11 matches with 1 you will earn 20 euros.
}
\examples{
library(data.table)
matches <- get_random_matches()
prob_voted_table <- calculate_probabilities(matches)
calculate_all_reward(prob_voted_table, 2600000)

}
