% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities.R
\name{calculate_probabilies}
\alias{calculate_probabilies}
\title{Calculate the prob_voted table}
\usage{
calculate_probabilies(matches, normalize = FALSE)
}
\arguments{
\item{matches}{(data.table): The real probability of each sign and the voted
percentage of each sign. Expected, at least, 6 columns named voted_1, voted_x,
voted_2, real_1, real_x and real_2.}

\item{normalize}{(boolean): TRUE if you want to apply a normalization over
the undervoted and overvoted signs.}
}
\value{
prob_voted_table (data.table): the table with prob and voted columns.
}
\description{
The prob_voted table is a big table that contains 4.7M rows with the possible
results and, for each row:
\itemize{
\item The probability of someone else to have voted with 14,13,12,11 and 10 matches right
if that row is the final result.
\item The probability of having 14,13,12,11 and 10 matches right if you vote this row.
}
}
\details{
For example, if the output of a row is voted_13 = 0.2 and prob_13 = 0.3, it means that:
\itemize{
\item For 10 external bets (other players bets), 2 of them will get 13 matches right.
\item You have 0.3 probabilities of having 13 matches right voting this row.
}
}
\examples{
calculate_probabilies(matches)

}
