% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_signs.R
\name{filter_possible_results}
\alias{filter_possible_results}
\title{Filter the prob_voted_table with possible final results}
\usage{
filter_possible_results(prob_voted_table, results)
}
\arguments{
\item{prob_voted_table}{(data.table): a table with a column sign that represents all bets to filter.}

\item{results}{(list of characters): A list with 14 characters with the known results,
'1','x' or '2' for each element. The unknown results must be represented with ''.}
}
\value{
possible_results (data.table): The prob_voted_table filtered with the final possible results
}
\description{
Given a set of matches which result is already known, it returns
the filtered prob_voted_table with only the final results that
are possible.
}
\examples{
library(data.table)
my_table <- data.table(sign = c('11111111111111', '1111111111111x'))
results <- c('','','','','', '', '', '','','','','','','x')
possible_results <- filter_possible_results(my_table, results)

}
