% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{get_data_eduardolosilla}
\alias{get_data_eduardolosilla}
\title{Scrapper function to get matches data from eduardolosilla's web}
\usage{
get_data_eduardolosilla(temporada, jornada)
}
\arguments{
\item{temporada}{(string): The season's bet. The season is always encoded with the bigger year, that means that
the season 2022/2023 is requested with temporada = 2023}

\item{jornada}{(string): The jornada's bet.}
}
\value{
matches (data.table): A table with the matches information
}
\description{
The function sends a request to eduardolosilla's web, parses the
response and returns the matches real and voted probabilities.
}
\examples{
\dontrun{
get_data_eduardolosilla('2023', '33')
}

}
