% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{get_data_quinielaticas}
\alias{get_data_quinielaticas}
\title{Scrapper function to get matches data}
\usage{
get_data_quinielaticas(url)
}
\arguments{
\item{url}{(string): The url where the html is located}
}
\value{
matches (data.table): A table with the matches information
}
\description{
The function reads the html placed in the param url
and returns the real and voted info for each match.
}
\details{
Note that the scrapper is only expecting the 'https://quinielaticas.com/' or
'https://quinielaticas.com/analisis-de-la-proxima-jornada-de-la-quiniela/' html
depends on if you prefer this week or next week matches.
}
\examples{
get_data_quinielaticas('https://quinielaticas.com/')

}
