% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_esperanza.R
\name{get_em}
\alias{get_em}
\title{Calculate EM for a given prob-voted table}
\usage{
get_em(prob_voted_table, money, level = 10)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table to work with. Watch calculate_probabilities
function for further information.}

\item{money}{(numeric): the total money played by all the bets}

\item{level}{(integer): The minimum EM to be calculted. For example,
with 'level = 12', EM14, EM13 and EM12 will be calculated. That is the
EM of getting 14, 13 and 12 right matches. Note that level must be an integer
between 10 and 14.}
}
\value{
prob_voted_table (data.table): The same prob_voted_table with em columns added.
}
\description{
EM is calculated as the money you would earn with an event multiplied by
the probability of that event to happen. The param level controls the
events to be calculated
}
\examples{
get_em(prob_voted_table, 2600000, 10)

}
