% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_esperanza.R
\name{get_group_cols}
\alias{get_group_cols}
\title{Calculate the group by columns for the get_EM function}
\usage{
get_group_cols(singles_signs, column_signs, out_columns)
}
\arguments{
\item{singles_signs}{(list string): the columns valid for group by process with just one sign}

\item{column_signs}{(list string): The columns valid for group by process}

\item{out_columns}{(list integer): The columns to exclude}
}
\value{
get_col (list string): The columns to perform the group by
}
\description{
This is a function used inside the get_EM function. It is not
useful for other purposes
}
