% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities.R
\name{get_prob_with_fails}
\alias{get_prob_with_fails}
\title{Transforms the prob_voted table splited by rows to the normal prob_voted table.}
\usage{
get_prob_with_fails(prob_voted_table_splited)
}
\arguments{
\item{prob_voted_table_splited}{(data.table): the prob_voted table splited by rows returned by
get_voted_prob function.}
}
\value{
prob_voted_table (data.table): the joined prob_voted table.
}
\description{
Watch get_probabilities and get_voted_prob functions if need further information
about the prob_voted table.
}
\examples{
get_prob_with_fails(table_prob_splited)
}
