% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_profitability.R
\name{get_profitability}
\alias{get_profitability}
\title{Calculate profitability for a given prob-voted table}
\usage{
get_profitability(prob_voted_table, importance = 1)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table to work with. Watch calculate_probabilities
function for further information.}

\item{importance}{(numeric): the importance given to the probability.
The recommended values are between 1 and 2. Default importance = 1}
}
\value{
prob_voted_table (data.table): The same prob_voted_table with profitability columns added.
}
\description{
The profitability is calculated as the real probability divided for the voted probability.
The profitability is used as a metric to order the bets. It is much faster than the em
but less exact, as it doesn't take into account that there are many bets that will have
10, 11, 12 and 13 signs right.
}
\details{
As the raw profitability gives too much importance to the voted probability (that will
end with a low real probability bet selection), a normalization parameter called importance was added.
Importance = 1 is the raw profitability.
}
\examples{
\dontrun{
library(data.table)
matches <- get_random_matches()
prob_voted_table <- calculate_probabilities(matches)
get_profitability(prob_voted_table, 1.5)
}

}
