% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matches.R
\name{get_random_matches}
\alias{get_random_matches}
\title{Function to get random matches data}
\usage{
get_random_matches(uniform = FALSE)
}
\arguments{
\item{uniform}{(boolean): If TRUE, all real and voted probabilities returned are
equal (0.33). Default value is FALSE.}
}
\value{
matches (data.table): A table with the random matches information
}
\description{
The function generates real and voted random probabilities
}
\examples{
get_random_matches()

}
