% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rentability.R
\name{get_rentability}
\alias{get_rentability}
\title{Calculate rentability for a given prob-voted table}
\usage{
get_rentability(prob_voted_table, importance = 1)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table to work with. Watch calculate_probabilities
function for further information.}

\item{importance}{(numeric): the importance given to the probability.
The recommended values are between 1 and 2. Default importance = 1}
}
\value{
prob_voted_table (data.table): The same prob_voted_table with rentability columns added.
}
\description{
The rentability is calculated as the real probability divided for the voted probability.
The rentability is used as a metric to order the bets. It is much faster than the em
but less exact, as it doesn't take into account that there are many bets that will have
10, 11, 12 and 13 signs right.
}
\details{
As the raw rentability gives too much importance to the voted probability (that will
end with a low real probability bet selection), a normalization parameter called importance was added.
Importance = 1 is the raw rentability.
}
\examples{
library(data.table)
matches <- get_random_matches()
prob_voted_table <- calculate_probabilities(matches)
get_rentability(prob_voted_table, 1.5)

}
