% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_probabilities.R
\name{increase_real_by_voted}
\alias{increase_real_by_voted}
\title{Fix matches real probability with voted probability}
\usage{
increase_real_by_voted(matches, index)
}
\arguments{
\item{matches}{(data.table): a table with 6 expected columns:
real_1, real_x, real_2, voted_1, voted_x and voted_2.}

\item{index}{(integer list): a list with 1 to 14 integers
that represent the indexes of the matches to be change.}
}
\value{
matches (data.table): the table in the input with the
real_1, real_x and real_2 changed with voted columns.
}
\description{
Some times the voted probability seams more reasonable than
real probability given by betting houses. To fix that you can
use this function.
}
\examples{
increase_real_by_voted(matches, c(4,7))

}
