% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_combination.R
\name{is_diff_sign}
\alias{is_diff_sign}
\title{Function to check if a sign has some distance conditions}
\usage{
is_diff_sign(data, selected, new, min_diff, total_diff)
}
\arguments{
\item{data}{(data.table): A table with expected column 'sign'}

\item{selected}{(list of integer): A list with the position in data of selected bets}

\item{new}{(integer): the position in data of the new bet}

\item{min_diff}{(integer): the min distance to check, sign by sign, between selected and new.}

\item{total_diff}{(integer): the min matches with different results when joined seleted and new.}
}
\value{
is_diff (boolean): Wheter if new has the two conditions or not
}
\description{
The function returns true if the new sign has the next two conditions:
\itemize{
\item At least a distance of min_diff with every sign in selected.
\item Joining with selected, at least there are total_diff matches with not all signs equal.
}
}
\examples{
library(data.table)
data <- data.table(sign = c('11111111111111', '11111111112222', '11122222111111'))
is_diff_sign(data, c(1), c(2), 5, 4)
is_diff_sign(data, c(1), c(2), 4, 4)

}
