% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_probabilities.R
\name{normalize_voted}
\alias{normalize_voted}
\title{Fix matches voted probability when voting too early}
\usage{
normalize_voted(matches)
}
\arguments{
\item{matches}{(data.table): a table with 6 expected columns:
real_1, real_x, real_2, voted_1, voted_x and voted_2.}
}
\value{
matches (data.table): the table in the input with the
voted columns normalized
}
\description{
Some times you cant vote the last day so voted probability
may change within the time. In order to try to reduce the impact
you can use this function. Note that this is a simple normalization,
it has no prediction component.
}
\examples{
normalized_voted(matches)

}
