% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bet.R
\name{read_bet}
\alias{read_bet}
\title{Read a file with bets}
\usage{
read_bet(path)
}
\arguments{
\item{path}{(string): The path where the file is located}
}
\value{
bets (data.table): A table with 1 column named sign
that contains the bets.
}
\description{
The file must be a .txt file with one bet per row.
Every bet must be composed by 14 characters with 1,2 or x
and 2 more with the 15th match, indicating 0,1,2 or M
}
\examples{
read_bet('file.txt')

A valid example of file.txt with 3 bets would be:
1111111111111102
1211111111111102
11111111111X11M2

}
