% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_signs.R
\name{signs_with_distance}
\alias{signs_with_distance}
\title{Calculate the signs that have n different matches with other given sign}
\usage{
signs_with_distance(out, this_id, dist, allow_lower_fails = FALSE, mysign)
}
\arguments{
\item{out}{(data.table): Ignored param, just used for recursive call in internal
logic.}

\item{this_id}{(integer): Ignored param, just used for recursive call in internal
logic.}

\item{dist}{(integer): The number of different sign between mysign and output.}

\item{allow_lower_fails}{(boolean): If true, returns the matches with lower different signs than dist.
Default set to false.}

\item{mysign}{(string): A string with 14 characters, '1', '2' or 'x' representing the sign
to calculate.}
}
\value{
return (data.table): A table with one column named sign with the signs that
have dist matches different.
}
\description{
This is an internal function useful for other complex calculations. It returns
all the signs which distance to other is n. For example, The signs that have
distance 1 with the sign '111' are
'x11','211','1x1','121','11x' and '112'
}
\details{
Note that sign must be of length 14
}
\examples{
signs_with_distance(dist = 3, mysign='11111111111111')

}
