% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_bet.R
\name{write_bet}
\alias{write_bet}
\title{Write a file with bets}
\usage{
write_bet(bet, path)
}
\arguments{
\item{bet}{(vector of strings): The list of bets to write}

\item{path}{(string): The path where the file will be located}
}
\description{
The file will be a .txt file with one bet per row.
}
\details{
A valid example of file.txt with 3 bets would be:
1111111111111102
1211111111111102
11111111111X11M2
}
\examples{
\dontrun{
bet = c('1111111111111102', '11111111111X11M2')
write_bet(bet, "file.txt")
}

}
