% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reward.R
\name{calculate_min_max_reward}
\alias{calculate_min_max_reward}
\title{Calculate the min and max reward for a given partial result and a set of bets}
\usage{
calculate_min_max_reward(prob_voted_table, played, results)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table with the expected reward
calculated}

\item{played}{(data.table): A column named sign with the character result is expected}

\item{results}{(list of characters): A list with 14 characters with the known results,
'1','x' or '2' for each element. The unknown results must be represented with ''.}
}
\value{
played (data.table): The same played table with min and max reward columns added.
}
\description{
Given a set of matches which result is already known and a set of bets,
it returns the max and min money you can earn. The max and min reward depends
on the result of the matches which result is unknown.
}
\examples{
results <- c('x','1','x','2','', '', '', '','x','2','1','','','')
min_max_rewards <- calculate_min_max_reward(prob_voted_table, selected, results)

}
