% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reward.R
\name{calculate_probability_of_reward}
\alias{calculate_probability_of_reward}
\title{Calculate the probability of been rewarded with more than an amount}
\usage{
calculate_probability_of_reward(
  prob_voted_table,
  played,
  results,
  rewards = c(10, 50, 100, 500, 1000, 5000)
)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table with the expected reward
calculated. Watch calculate_all_reward function for more information about the reward column.}

\item{played}{(data.table): A column named sign with the character result is expected}

\item{results}{(list of characters): A list with 14 characters with the known results,
'1','x' or '2' for each element. The unknown results must be represented with ''.}

\item{rewards}{(list of numeric): The list with the rewards whose probability is going to
be calculated. Default value c(10,50,100,500,1000,5000).}
}
\value{
prob_of_reward (data.table): A table with two columns, the rewards and the probability
of earning more of that rewards.
}
\description{
Given a set of matches which result is already known, a set of bets
and a list of amounts, the function returns the probability of been
rewarded with more than every amount.
}
\examples{
results <- c('x','1','x','2','', '', '', '','x','2','1','','','')
prob_of_reward <- calculate_probability_of_reward(prob_voted_table, played, results, rewards = c(10,20))

}
