% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reward.R
\name{calculate_probability_of_roi}
\alias{calculate_probability_of_roi}
\title{Calculate the probability of been rewarded with more than the invested amount}
\usage{
calculate_probability_of_roi(prob_voted_table, played, results, money)
}
\arguments{
\item{prob_voted_table}{(data.table): the prob_voted table to work with. Watch calculate_probabilities
function for further information.}

\item{played}{(data.table): A column named sign with the character result is expected}

\item{results}{(list of characters): A list with 14 characters with the known results,
'1','x' or '2' for each element. The unknown results must be represented with ''.}

\item{money}{(numeric): the total money played by all the bets}
}
\value{
prob_of_reward (data.table): A table with two columns, the rewards and the probability
of earning more of the invested amount (number of bets * 0.75).
}
\description{
Given a set of matches which result is already known, a set of bets,
the function returns the probability of been
rewarded with more than the invested amount.
}
\examples{
\dontrun{
library(data.table)
matches <- get_random_matches()
prob_voted_table <- calculate_probabilities(matches)
played <- data.table(sign = c('11111111111111', '11211111111111'))
results <- c('x','1','x','2','', '', '', '','x','2','1','','','')
money <- 1000000
prob_of_reward <- calculate_probability_of_roi(prob_voted_table, played, results, money)
}

}
