% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_signs.R
\name{filter_table_prob}
\alias{filter_table_prob}
\title{Filter the prob_voted_table with specified results in matches}
\usage{
filter_table_prob(prob_voted_table, matches, results)
}
\arguments{
\item{prob_voted_table}{(data.table): a table with a column sign that represents all bets to filter.}

\item{matches}{(list of integers): A list with n integer numbers where the results are applied.}

\item{results}{(list of characters): A list with n signs to filter.}
}
\value{
possible_results (data.table): The prob_voted_table filtered with the specified results in matches
}
\description{
Given a set of matches which result is already known, it returns
the filtered prob_voted_table with only the results that
satisfy the specified signs.
}
\examples{
library(data.table)
my_table <- data.table(sign = c('x1111111111111', '11111111111111'))
filtered_table <- filter_table_prob(my_table, c(1), 'x')

}
