% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities.R
\name{get_prob_with_fails}
\alias{get_prob_with_fails}
\title{Transforms the prob_voted table splited by rows to the normal prob_voted table.}
\usage{
get_prob_with_fails(prob_voted_table_splited)
}
\arguments{
\item{prob_voted_table_splited}{(data.table): the prob_voted table splited by rows returned by
get_voted_prob function.}
}
\value{
prob_voted_table (data.table): the joined prob_voted table.
}
\description{
Watch get_probabilities and get_voted_prob functions if need further information
about the prob_voted table.
}
\examples{
library(data.table)
matches <- get_random_matches()
matches[,':='(
  voted_no_1 = 1 - voted_1, voted_no_x = 1 - voted_x, voted_no_2 = 1 - voted_2,
  real_no_1 = 1 - real_1, real_no_x = 1 - real_x, real_no_2 = 1 - real_2
)]
prob_voted_table_splited = get_voted_prob(matches)
get_prob_with_fails(prob_voted_table_splited)

}
