% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_probabilities.R
\name{get_voted_prob}
\alias{get_voted_prob}
\title{Calculate the prob_voted table splited by rows}
\usage{
get_voted_prob(matches, out = data.table(), this_id = 1, allow_fails = 4)
}
\arguments{
\item{matches}{(data.table): The matches to calculated the probabilities.}

\item{out}{(data.table): Ignored param, just used for recursive call in internal
logic.}

\item{this_id}{(integer): Ignored param, just used for recursive call in internal
logic.}

\item{allow_fails}{(integer): The number of max fails you want to work with. For example,
a value of 2 will return the only the prob and voted of 0,1 and 2 wrong matches.
Default value is 4}
}
\value{
prob_voted_table_splited (data.table): the prob_voted table splited by rows.
}
\description{
The prob_voted table splited by rows is similar to the prob_voted table but,
instead of having columns for 14,13,12,11 and 10 matches right, you only have
four columns: the matches results, the prob, the voted and the fails you expect.
}
\details{
That way, you will have 4.7M * (n+1) rows,
with n the number of allow_fails (typically 4)
}
\examples{
library(data.table)
matches <- get_random_matches()
matches[,':='(
  voted_no_1 = 1 - voted_1, voted_no_x = 1 - voted_x, voted_no_2 = 1 - voted_2,
  real_no_1 = 1 - real_1, real_no_x = 1 - real_x, real_no_2 = 1 - real_2
)]
get_voted_prob(matches)

}
